210409689 - Arreglado un fallo crítico en el ACCESS de ChanServ.
git-svn-id: file:///cygdrive/c/CTTC/backup/redyc/svn/Colossus/trunk@766 8e1883df-7ee1-ef4f-9e60-6d04f2a2cf0e

+++ src/modulos/chanserv.c
@@ -1610,6 +1610,16 @@ BOTFUNC(CSAccess)
 	else if (params == 3)
 	{
 		char *n_c, *c_c;
+		if (!CSTieneNivel(cl, param[1], CS_LEV_EDT))
+		{
+			Responde(cl, CLI(chanserv), CS_ERR_FORB, "");
+			return 1;
+		}
+		if (!CSEsAccess(param[1], param[2]))
+		{
+			Responde(cl, CLI(chanserv), CS_ERR_EMPT, "Este nick no tiene acceso.");
+			return 1;
+		}
 		c_c = SQLEscapa(param[1]);
 		n_c = SQLEscapa(param[2]);
 		SQLQuery("DELETE FROM %s%s WHERE canal='%s' AND nick='%s'", PREFIJO, CS_ACCESS, c_c, n_c);
